Revert "merge main and fix conflicts"
This reverts commit 4f13358e3f9d487f4bb8f2e4ecb58cd8c2b4b357, reversing
changes made to 644bc7857b8e3a80167e953671f51deda4e231d9.

+++ impl/camb/common/transpose.cpp
@@ -66,7 +66,7 @@ diopiError_t transpose(diopiContextHandle_t ctx, const DiopiTensor& inputTensor,
     DIOPI_CALLCNNL(cnnlSetTransposeDescriptor(cnnlTransposeDesc.get(), perms.size(), perms.data()));
     if (!outTensor.defined()) {
         std::vector<int64_t> trShape(perms.size());
-        for (size_t i = 0; i < perms.size(); ++i) {
+        for (int i = 0; i < perms.size(); ++i) {
             trShape[i] = inputTensor.shape()[perms[i]];
         }
         outTensor = requiresTensor(ctx, trShape, inputTensor.dtype());
